+++ CHANGELOG.md
@@ -7,6 +7,14 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+* [PR-12](https://github.com/rimi-itk/dpl_pretix/pull/12)
+  More fixes after test:
+
+  * Handled ticket URL on event instance form
+  * Set ticket URL on event instances
+  * Enforced ticket category
+  * Included place in location
+
 * [PR-11](https://github.com/rimi-itk/dpl_pretix/pull/11)
   Fixed some issues
 * [PR-10](https://github.com/rimi-itk/dpl_pretix/pull/10)